% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_funs.R
\name{bisection}
\alias{bisection}
\title{Using bisection method to find the iv}
\usage{
bisection(cm, k, cp = 1, vl, vh)
}
\arguments{
\item{cm}{option market mid price}

\item{k}{log(K/F)}

\item{cp}{call/put flag (call 1, put -1)}

\item{vl}{lower bound of total IV}

\item{vh}{upper bound of total IV}
}
\description{
Vectoring IV estimation with the bisection method
}
