% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_funs.R
\name{bsPrice}
\alias{bsPrice}
\title{Calculate BS option price}
\usage{
bsPrice(forward, strike, tau, iv, cp = 1, R = 0.02)
}
\arguments{
\item{forward}{Forward price}

\item{strike}{strike}

\item{tau}{maturity}

\item{iv}{implied volatility}

\item{cp}{call/put flag (call 1, put -1)}

\item{R}{interest rate}
}
\description{
Calculate BS option price
}
