% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calibration.R
\name{cal.price}
\alias{cal.price}
\title{estimate calendar prices}
\usage{
cal.price(surf, dte, strike, positions = c(-1, 1), spot = NULL,
  type = "put", trueDelta = TRUE)
}
\arguments{
\item{surf}{a given surface}

\item{dte}{Days to expiration}

\item{positions}{positions of the strikes}

\item{strikes}{strikes of the calendar}
}
\description{
Estimate calendar combo price, delta, theta based on a given
surface.
}
