% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr_calibration.R
\name{calibrate.sabr}
\alias{calibrate.sabr}
\title{Calibrate SABR}
\usage{
calibrate.sabr(df, b = 0.5)
}
\arguments{
\item{df}{a data.frame containing strikes, ivs, forwards, tau}

\item{b}{beta}
}
\description{
calibrate SABR model and return the SABR parameters.
}
