% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr_calibration.R
\name{calibrate.sabrvol}
\alias{calibrate.sabrvol}
\title{Calibrate SABR model}
\usage{
calibrate.sabrvol(fwd, strikes, ivs, t, b, v0)
}
\arguments{
\item{fwd}{forward}

\item{strikes}{strikes}

\item{ivs}{implied volatilities}

\item{t}{tau}

\item{b}{beta parameter from SABR}
}
\description{
Calibrate SABR model. Note the parameters t and b must be scalar.
}
