% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{chain.iv}
\alias{chain.iv}
\title{Calculate chain iv, F, U}
\usage{
chain.iv(chain, LR = 0.01)
}
\arguments{
\item{chain}{call/put paired option chain}

\item{LR}{liborrate}
}
\description{
Base on BS model
}
