% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{chainForwardPrice}
\alias{chainForwardPrice}
\title{estimate forward price with libor rate}
\usage{
chainForwardPrice(pairChain, LR = 0.0169)
}
\description{
calculate forward price for the CP-pair chain with known interest rates
Input: paired chain, interest rate
}
