% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{estimate.iv}
\alias{estimate.iv}
\title{Estimate chain ivs, F, U}
\usage{
estimate.iv(chain, LR = 0.01)
}
\arguments{
\item{chain}{call/put paired option chain}

\item{LR}{liborrate}
}
\description{
Using bisection method and tighter bounds proposed by Gatheral et al. 2017
}
