% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr_calibration.R
\name{fit.termstructure}
\alias{fit.termstructure}
\title{Fit Term structure over tau}
\usage{
fit.termstructure(sdf, tr = c(14, 180), b = 0.5)
}
\arguments{
\item{sdf}{data.frame of sabr model}
}
\description{
fit non-linear relation of v0(t), rho(t), and volvol(t)
}
