% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{fit_ga_svi}
\alias{fit_ga_svi}
\title{Estimate svi parameers with GA algorithm}
\usage{
fit_ga_svi(k, iv, tau)
}
\arguments{
\item{k}{log moneyness}

\item{iv}{iv}

\item{tau}{tau}
}
\description{
This require ga package, GA may give a robust result
}
