% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{fit_ga_svih}
\alias{fit_ga_svih}
\title{estimate svi parameter with know exponent H}
\usage{
fit_ga_svih(k, iv, tau, H)
}
\arguments{
\item{k}{log moneyness}

\item{iv}{iv}

\item{tau}{tau}
}
\description{
Using GA method to estimate parameters
}
