% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{fit_inner}
\alias{fit_inner}
\title{Inner optimization for parameter: c,d,a}
\usage{
fit_inner(y, w, sig, H)
}
\arguments{
\item{y}{(k-m)/sigma}

\item{w}{real time-weighted iv data, usually iv*sqrt(tau)}

\item{sig}{sigma}

\item{H}{fractional exponent}
}
\value{
a,b,rho parameters
}
\description{
optimization of the inner parameters
let y(k) = (k - m)/sig
svi w = a + b*sig*(rho*y(k) + sqrt(y(x)^2+1))
       = a + dy(x) + cz(x)
d = roh*b*sig, c = b*sig, z = sqrt(y(x)^2+1)
Constrains:
    0 <= c <= 4sig
    |d| <= c
    |d| <= 4sig -c
    0 <= a <= max_i(w_i)
}
