% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{fit_quasi_svi}
\alias{fit_quasi_svi}
\title{fit modifed SVI function}
\usage{
fit_quasi_svi(k, iv, tau)
}
\arguments{
\item{k}{log moneyness}

\item{iv}{iv}

\item{tau}{tau}
}
\value{
svi 6 parameters
}
\description{
The function will fit the 6 parameters, including exponent H, for the SVI function
Due to the exponent H, the fitting time is slow. It need to be improved.
}
