% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{fit_quasi_svih}
\alias{fit_quasi_svih}
\title{Estimated svi parameters with a given H}
\usage{
fit_quasi_svih(k, iv, tau, H)
}
\arguments{
\item{k}{log moneyness}

\item{iv}{iv}

\item{tau}{tau}

\item{H}{exponent H for the modified SVI function}
}
\description{
Estimate modified svi parameters with a given H.
}
