% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{fit_raw_surface}
\alias{fit_raw_surface}
\title{fit surface with all vol}
\usage{
fit_raw_surface(implied_volatility, maturity, log_moneyness,
  phifun = "power_law")
}
\description{
Inputs:
   implied volatility
   maturity
   log moneyness
   phifun
}
