% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{fit_ssvi}
\alias{fit_ssvi}
\title{Fit SSVI}
\usage{
fit_ssvi(phifun = "power_law", log_moneyness, theta_expanded,
  total_implied_variance)
}
\description{
Inputs:
phifun = default 'power_law'
log_moneyness
theta
total implied variance
}
