% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{fit_svi}
\alias{fit_svi}
\title{FIT SVI}
\usage{
fit_svi(x0, k, total_implied_variance, slice_before, slice_after, tau)
}
\description{
fit_svi fits a volatility slice to observed implied volatilities.
Input:
  x0, k, total implied variance, slice before, slice after, tau
Output:
  estimated surface parameters
}
