% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{fit_svi_surface}
\alias{fit_svi_surface}
\title{Fit SVI surface}
\usage{
fit_svi_surface(implied_volatility, maturity, log_moneyness,
  phifun = "power_law")
}
\description{
fit_svi_surface calibrates the SVI surface to market data. First, the entire Surface SVI is fitted
to all log-moneyness-theta observations. Second, each slice is fitted again using the SSVI fit as
initial guess.
Input:
* impliedvolatility, maturity, moneyness, phi
Output:
* parameters = (2x1) = parameters of SSVI = [rho, lambda]
* theta      = (Tx1) = ATM total implied variance
* maturities = (Tx1) = corresponding time to maturity
}
