% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{fitfunction1}
\alias{fitfunction1}
\title{fit function for GA optimization}
\usage{
fitfunction1(x, phifun = "power_law", log_moneyness, theta,
  total_implied_variance)
}
\description{
fit function for GA optimization
}
