% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{fitfunction2}
\alias{fitfunction2}
\title{The objective function of the minimization of fit_svi.}
\usage{
fitfunction2(x, k, total_implied_variance, slice_before, slice_after, tau)
}
\description{
The objective is the 2-norm
of the error in total implied variance. If specified, the function tests whether the current total
variance slice lies between the prior and later slice. If the arbitrage bound is violated, the
function is penalized with a very large value.
Input:
* x = (5x1) = parameters of SVI model
* k = (Nx1) = log moneyness at which total implied variance is calculated
* total_implied_variance = (Nx1) = market observed total implied variance
* slice_before = (Nx1) = total implied variance with smaller maturity than current slice
* slice_after = (Nx1) = total implied variance with larter maturity than current slice
}
