% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{fitfunction3}
\alias{fitfunction3}
\title{fit function for initial optimization}
\usage{
fitfunction3(x, phifun = "power_law", log_moneyness, theta,
  total_implied_variance)
}
\description{
fit function for initial optimization
}
