% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calibration.R
\name{fly.price}
\alias{fly.price}
\title{estimate combo prices}
\usage{
fly.price(surf, dte, strikes, positions = c(1, -2, 1), spot = NULL,
  type = "put", trueDelta = TRUE)
}
\arguments{
\item{surf}{a given surface}

\item{dte}{Days to expiration}

\item{strikes}{strikes of the fly}

\item{positions}{positions of the strikes}
}
\description{
Estimate combo (e.g., butterfly) price, delta, theta based on a given
surface.
}
