% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_funs.R
\name{l_bs}
\alias{l_bs}
\title{BS total IV approximation (tight lower bound)}
\usage{
l_bs(c, k, cp = 1)
}
\arguments{
\item{c}{normalized option price C/D/F (call), or P/D/F (put). D = e^(-rt) discount, F forward price}

\item{k}{log(K/F)}

\item{cp}{call/put indicator, call = 1, put = -1}
}
\description{
Black-Scholes implied volatility approximation, adopted from Stefanica and Radoicic (2017)
An explicit implied volatility formula, the approximation error < 2%, maximum error < 10%
}
