% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{livevol.chain}
\alias{livevol.chain}
\title{Import livevol zip file and generate chain and surface}
\usage{
livevol.chain(filename, rate = 0.017)
}
\arguments{
\item{filename}{filename of an input zip file}

\item{rate}{Interest rate at that day}
}
\value{
a list contain the chain and surface, and save two to files
}
\description{
Import livevol zip file and generate chain and surface
}
