% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{optTransactions}
\alias{optTransactions}
\title{select option transactions from trade log table}
\usage{
optTransactions(op_df, sym)
}
\arguments{
\item{op_df}{option table}

\item{symbol}{target symbol, e.g., 'SPX'}
}
\value{
option transaction with expiration, strike, callput, ...
}
\description{
select option transactions from trade log table
}
