% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr_calibration.R
\name{plot.sabrfit}
\alias{plot.sabrfit}
\title{Visualize SABR fit}
\usage{
\method{plot}{sabrfit}(sdf, chain, tr = c(0.1, 1))
}
\arguments{
\item{sdf}{data.frame of sabr model}

\item{chain}{raw chain}

\item{rt}{range of tau}
}
\description{
Visualize SABR goodness of fit
}
