% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_stiching.R
\name{plot.stiching_surface}
\alias{plot.stiching_surface}
\title{plot iv surface with fitted model}
\usage{
\method{plot}{stiching_surface}(ivchain, para, range = c(7, 90),
  type = "iv", x = "strike")
}
\arguments{
\item{ivchain}{iv chain}

\item{para}{svi parameters}

\item{range}{dte range (default c(7,90))}

\item{type}{default ('iv') plot normal IV figure, or 'ivt', plot iv/sqrt(t), or 'var', plot total}

\item{x}{default ('strike'), alternative 'logmoneyness'
implied variance (i.e., iv^2*tau)}
}
\description{
plot iv surface with fitted model
}
