% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr_calibration.R
\name{plot.surface}
\alias{plot.surface}
\title{volatility surface}
\usage{
\method{plot}{surface}(sdf, kr = c(0.7, 1.3), tr = c(10, 180))
}
\arguments{
\item{sdf}{data.frame of sabr parameters}

\item{kr}{relative strike k range}

\item{tr}{tau range}
}
\description{
Volatility surface based on SABR model
}
