% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr_calibration.R
\name{plot.termstructure}
\alias{plot.termstructure}
\title{Term structure of vol surface}
\usage{
\method{plot}{termstructure}(sdf, tr = c(14, 180))
}
\arguments{
\item{sdf}{data.frame of sabr model}
}
\description{
Visualize the term structure (Vol ATM) from SABR model
}
