% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{quasi.err}
\alias{quasi.err}
\title{estimate surface error}
\usage{
quasi.err(ivchain, para, range = c(7, 90), show = FALSE)
}
\description{
Provide a chain and surface parameters, show the error
By default, the graph output is off.
}
