% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{quasi.estimateSurface}
\alias{quasi.estimateSurface}
\title{estimate surface from a chain file}
\usage{
quasi.estimateSurface(filename, range = c(7, 90))
}
\description{
Estimate surface with modified SVI model using a chain file
The chain file name should be 'rds'. This function uses parallel
computing.
}
\details{
Return the whole chain.
}
