% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{quasi.interIV}
\alias{quasi.interIV}
\title{interpolation of IVs with surface}
\usage{
quasi.interIV(surf, k, tau_interp)
}
\arguments{
\item{surf}{surface parameters}

\item{k}{logmoneyness}

\item{tau_interp}{to-be-interpolated tau}
}
\value{
interpoplated IVs at tau_interp
Note: Interpolation works only within the maturity range.
}
\description{
interpolation of IVs with surface
}
