% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{quasi.ivs}
\alias{quasi.ivs}
\title{estimate iv curve and skew at a given DTE using modified SVI function}
\usage{
quasi.ivs(surface, dte, spot = NULL, type = "put", x = seq(-0.2, 0.1,
  length.out = 600))
}
\arguments{
\item{surface}{a surface param, estimated from SVI}

\item{dte}{Days to expiration}

\item{x}{a sequence of logmoneyness or strike. If max(x)<1, it will be regarded as log moneyness.
It return a table with iv, skew}

\item{n}{a smoothing factor for skew curve}
}
\description{
This function will estimate a given iv curve with a specified dte
}
