% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{quasi.skew}
\alias{quasi.skew}
\title{estimate skews across DTES}
\usage{
quasi.skew(surface, dtes = c(10, 20, 30, 40, 50, 60, 70), x = seq(-0.1, 0.1,
  length.out = 300), plot = TRUE)
}
\arguments{
\item{surface}{svi surface parameters}

\item{dtes}{dtes}

\item{x}{log moneyness or strikes}
}
\description{
estimate skews from the modified SVI surface
}
