% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{quasi.surface.at}
\alias{quasi.surface.at}
\title{plot surface with a given hour and minutes in a day}
\usage{
quasi.surface.at(chain, surf, hourtime, type = "ivt", range = c(7, 100))
}
\arguments{
\item{chain}{chains}

\item{hourtime}{time, e.g., "10:30"}
}
\description{
This function to plot furface with a given hour and minutes, such as "10:30".
It will then extract the chain and surface and plot it.
}
