% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{quasi.termstructure}
\alias{quasi.termstructure}
\title{Plot term structure}
\usage{
quasi.termstructure(surface, plot = TRUE)
}
\arguments{
\item{surface}{surface parameters}

\item{plot}{plot the graph or not}
}
\description{
Based on modified svi surface, estimate ATM IVs across DTEs
}
