% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{read.csv.zip}
\alias{read.csv.zip}
\title{Read csv data from zip file}
\usage{
read.csv.zip(zipfile, pattern = "\\\\.csv$", ...)
}
\arguments{
\item{zipfile}{filename of the zip file}

\item{pattern}{the pattern that match the file name, such as '.csv'}
}
\description{
Read a large csv file from a zip file
If the zip file contains multiple csv files, all files will be read and concatenated.
}
