% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{read.ivolchain}
\alias{read.ivolchain}
\title{Read ivolatility option chain
read option chain raw data downloaded from ivolatility
read ivolatility cvs file.}
\usage{
read.ivolchain(filename)
}
\description{
Read ivolatility option chain
read option chain raw data downloaded from ivolatility
read ivolatility cvs file.
}
