% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{read.livevol}
\alias{read.livevol}
\title{Read livevol zip csv chain}
\usage{
read.livevol(filename, LR = 0.017)
}
\description{
Import csv raw chain, and calculate iv based on bid/ask
return a refined chain
}
