% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{read.vuechain}
\alias{read.vuechain}
\title{Read optionVue option chain file}
\usage{
read.vuechain(filename)
}
\description{
read optionVue exported cvs file. It should contain:
Th.Price, Date, Exp Date, Strike Price,
Mid IV, Call/Put, Item Type, Market Price
}
