% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{readTradeLog}
\alias{readTradeLog}
\title{Read trade log text file}
\usage{
readTradeLog(logfile)
}
\value{
stock transaction,
   option transaction,
   stock position,
   option position
   to the global environment
}
\description{
Read trade log text file
}
