% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr_calibration.R
\name{sabrVol}
\alias{sabrVol}
\title{Log-normal vol by SABR}
\usage{
sabrVol(f, k, t, alpha, beta, rho, volvol)
}
\arguments{
\item{f}{Forward}

\item{k}{strike}

\item{alpha}{alpha parameter}

\item{beta}{beta parameter}

\item{rho}{rho parameter}

\item{volvol}{volvol parameter}
}
\description{
Calculate log-normal vol based on sabr model parameters
}
