% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_stiching.R
\name{stiching.curve}
\alias{stiching.curve}
\title{estimate iv curve and skew at a given DTE}
\usage{
stiching.curve(param, dte, x = seq(-0.2, 0.1, length.out = 600))
}
\arguments{
\item{param}{a surface param, estimated from SVI}

\item{dte}{Days to expiration}

\item{x}{a sequence of logmoneyness or strike. If max(x)<1, it will be regarded as log moneyness.
It return a table with iv, skew}

\item{n}{a smoothing factor for skew curve}
}
\description{
This function will estimate a given iv curve with a specified dte
}
