% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_stiching.R
\name{stiching.option.History}
\alias{stiching.option.History}
\title{obtain optio price history from the chain}
\usage{
stiching.option.History(start_date, exp_date, strike_combo, wholechain)
}
\description{
If the chain or strike is not available, based on svi interpolation
Input: start date, expiration date, strike combo, whole option chain
}
