% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_stiching.R
\name{stiching.optionPrice}
\alias{stiching.optionPrice}
\title{option price based on svi-estimated iv surface}
\usage{
stiching.optionPrice(strike, param, tau, s = NULL, type = "put")
}
\description{
Inputs: type: 'call' or 'put'
        strike, s(pot), tau (maturity),
        param: iv surface parameter data.frame
Return: EuropeanOption (value, delta, gamma, vega, theta, rho, divRho)
§
}
\details{
optionPrice is now accept array of strike, spot, and type. Three must be the same length
}
