% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_stiching.R
\name{stiching.plcurves}
\alias{stiching.plcurves}
\title{Estimate P/L curves}
\usage{
stiching.plcurves(start, exp_date, combo, prange, chain, t = 0)
}
\arguments{
\item{start}{The intial date}

\item{exp_date}{expiration date}

\item{combo}{a combo option strategy, including three components:
strike, position size, and type (call/put)}

\item{prange}{price range that to be evaluated}

\item{chain}{option chain of the intial date}

\item{t}{days from the intial date >=0.}
}
\description{
Using obtained svi surface to estimate pl curves
}
