% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_stiching.R
\name{stiching.rnd}
\alias{stiching.rnd}
\title{Risk neutral probability based on the iv}
\usage{
stiching.rnd(param, dte, spot = NULL, type = "put", x = seq(-0.2, 0.1,
  length.out = 600))
}
\arguments{
\item{param}{svi surfance paramters}

\item{dte}{Days to expiration}

\item{type}{put or call}

\item{x}{x-axis (log moneyness or strike)}

\item{sopt}{default is the spot price of the surface that obtained. It can be different.}

\item{n}{smoothing factor for skew and pdf}
}
\description{
This function estimated a risk neutral probability /density function at a given DTE
}
