% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi.fitparam}
\alias{svi.fitparam}
\title{Fit Chain SVI parameters}
\usage{
svi.fitparam(chain)
}
\arguments{
\item{chain}{an iv option chain containing iv, maturity, logmoneyness
Note: chain should be from one single day}
}
\value{
surface parameter
}
\description{
Fit iv surface using svi approach with a given day chain
}
