% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi.hybridparam}
\alias{svi.hybridparam}
\title{Fit Chain surface with hybrid svi model}
\usage{
svi.hybridparam(chain, boundary = -0.02)
}
\arguments{
\item{chain}{an iv option chain containing iv, maturity, logmoneyness
Note: chain should be from one single day}
}
\value{
surface parameter
}
\description{
Fit iv surface using svi approach with a given day chain,
left side with a quadra-linear model, right side: svi
}
