% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi.interparam}
\alias{svi.interparam}
\title{interpolation of svi parameters}
\usage{
svi.interparam(para, tau_interp)
}
\arguments{
\item{para}{surface parameters}

\item{tau_interp}{to-be-interpolated tau}
}
\value{
interpoplated parameter at tau_interp
Note: Interpolation works only within the maturity range.
In this version, interpolation also for the linear part

values contains 5 of svi parameters, 3 of linear model
}
\description{
interpolation of svi parameters
}
