% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{svi.quasiparam.p}
\alias{svi.quasiparam.p}
\title{Fit Chain SVI parameters with quasi-calibration approach}
\usage{
svi.quasiparam.p(chain, H = NULL)
}
\arguments{
\item{chain}{an iv option chain containing iv, maturity, logmoneyness
Note: chain should be from one single day}
}
\value{
surface parameter (a,b,m, rho, sigma, H)
}
\description{
Fit iv surface using svi approach with a given day chain. Due to slowness, the function
first calculate H for one-month chain, and use this H for the whole surface.
This version uses parallel computing.
}
