% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi}
\alias{svi}
\title{Stochastic Volatility Inspired parameterization}
\usage{
svi(k, param, type, tau = 1)
}
\arguments{
\item{k}{= (Kx1) = moneyness at which to evaluate the surface}

\item{param}{= parameters of the surface}

\item{type}{= 'raw', 'natural','jumpwing'}

\item{tau}{expiration}
}
\description{
SVI - Stochastic Volatility Inspired parameterization of the implied
volatility smile. This function implements combination of three types formulation.
}
