% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi_convertparameters}
\alias{svi_convertparameters}
\title{SVI convert parameters}
\usage{
svi_convertparameters(param_old, from, to, tau = 1)
}
\arguments{
\item{param_old}{(5x1) original parameters}

\item{from}{string = formulation of original parameters (raw, natural, jumpwing)}

\item{to}{string = formulation of new parameters (raw, natural, jumpwings)}
}
\value{
param_new = (5x1) = new parameters
}
\description{
svi_convertparameters converts the parameter set of one type of SVI
formulation to another. The parameterizations are assumed to be:
* raw =(a,b,m,rho, sigma)
* natural = (delta, mu, rho, omega, zeta)
* jumpwing = (v, psi, p, c, vt)
}
