% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-calibration.R
\name{svi_fun}
\alias{svi_fun}
\title{svi curve function}
\usage{
svi_fun(k, param)
}
\arguments{
\item{k}{log moneyness}

\item{param}{6 parameters: a, b,m, rho, sigma, H}
}
\value{
estimated values according to the curve
}
\description{
Modified SVI function with fraction exponent H
modified version: svi^H, additional parameter
}
