% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi_iv_interp}
\alias{svi_iv_interp}
\title{A simplified version of svi_interpolation}
\usage{
svi_iv_interp(log_moneyness, tau_interp = 1, para)
}
\description{
Output:
* implied_volatility (Kx1) = implied volatilities corresponding to total_implied_variance
para is the output of the fitted surface function
}
