% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi_jumpwing}
\alias{svi_jumpwing}
\title{SVI Jump wing formulation}
\usage{
svi_jumpwing(k, param, tau = 1)
}
\arguments{
\item{k}{= (Nx1) = log-moneyness at which to evaluate the total implied
variance}

\item{param}{= (5x1) = parameters:
* v = scalar = ATM variance
* psi = scalar = ATM skew
* p = scalar = slope of left/put wing
* c = scalar = slope of right/call wing
* vt = scalar = minimum implied variance}
}
\value{
* totalvariance = (Nx1) = estimated total variance at k
* impliedvolatility = (Nx1) = estimated implied volatility at (k,tau)
}
\description{
SVI - Stochastic Volatility Inspired parameterization of the implied
volatility smile. This function implements the jump-wings formulation.
}
