% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi_natural}
\alias{svi_natural}
\title{SVI Natural formulation}
\usage{
svi_natural(k, param, tau = 1)
}
\arguments{
\item{k}{(Nx1) log-moneyness at which to evaluate the total implied
variance}

\item{param}{= (5x1) = parameters:
* delta = scalar = level of variance
* mu = scalar = slope of wings
* rho = scalar = translates smile to right,
* omega = scalar = counter-clockwise rotation of smile
* zeta = scalar = reduces ATM curvature of the smile}
}
\value{
* totalvariance = (Nx1) = estimated total variance at k
* impliedvolatility = (Nx1) = estimated implied volatility at (k,tau)
}
\description{
SVI - Stochastic Volatility Inspired parameterization of the implied
volatility smile. This function implements the natural formulation.
}
