% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi_parameter_bounds}
\alias{svi_parameter_bounds}
\title{SVI parameter bounds}
\usage{
svi_parameter_bounds(parameterization)
}
\description{
svi_parameter_bounds returns the parameter bounds of an SVI parameterization. The parameters are
assumed to be in the following order:
raw = [a b m rho sigma]
natural = [delta mu rho omega zeta]
jumpwing = [v psi p c vt]
}
