% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi_raw}
\alias{svi_raw}
\title{SVI Raw formulation}
\usage{
svi_raw(k, param, tau = 1)
}
\arguments{
\item{k}{log-moneyness at which to evaluate the total implied variance}

\item{(5x1)}{parameters:
  a = scalar = level of variance
  b = scalar = slope of wings
  m = scalar = translates smile to right
  rho = scalar = counter-clockwise rotation of smile
sigma = scalar = reduces ATM curvature of the smile}
}
\value{
A list contain
   totalvariance = (Nx1) = estimated total variance at k
   impliedvolatility = (Nx1) = estimated implied volatility at (k,tau)
}
\description{
SVI Raw formulation
}
