% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_funs.R
\name{svi_surface}
\alias{svi_surface}
\title{SVI Surface}
\usage{
svi_surface(k, theta, rho, phifun, phi_param, tau = 1)
}
\arguments{
\item{k}{= (Kx1) = moneyness at which to evaluate the surface}

\item{theta}{= (Tx1) = ATM variance time at which to evaluate the surface.
If theta and k have the same dimensions, then the surface is evaluated
for each given (k,theta) pair. If the dimensions are different, the
function evaluates the surface on the grid given by k and theta}

\item{param}{= (Xx1) = parameters of the surface}
}
\value{
* totalvariance = (KxT) or (Kx1) = estimated total variance
* impliedvolatility = (KxT) or (Kx1) = estimated implied volatility
}
\description{
svi_surface calcualtes the surface SVI free of statis arbitrage.
}
