% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calibration.R
\name{trans.fit_ab}
\alias{trans.fit_ab}
\title{Inner optimzation of the SVI function with known rho}
\usage{
trans.fit_ab(x, k, w)
}
\arguments{
\item{x}{two parameters (m, rho,sigma)}

\item{k}{log moneyness or alike}

\item{w}{total implied vol, in the format of iv^(1/H). When H = 1/2, it is normal total variance}
}
\value{
estimated 5 SVI parameters (a, b, m, rho, sigma)
}
\description{
Here given m and sigma and rho (in x), k, and total iv w, estimate a, b
}
