% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_calibration.R
\name{trans.fit_svi}
\alias{trans.fit_svi}
\title{fit SVI function
Here we assume iv^H = SVI_Fun(Param)
The function will fit the 6 parameters, including exponent H, for the SVI function}
\usage{
trans.fit_svi(k, iv)
}
\arguments{
\item{k}{log moneyness}

\item{iv}{iv}
}
\value{
svi 6 parameters
}
\description{
fit SVI function
Here we assume iv^H = SVI_Fun(Param)
The function will fit the 6 parameters, including exponent H, for the SVI function
}
