% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calibration.R
\name{trans.svi_fun}
\alias{trans.svi_fun}
\title{svi original function}
\usage{
trans.svi_fun(k, param)
}
\arguments{
\item{k}{log moneyness}

\item{param}{5 parameters: a, b,m, rho, sigma}
}
\value{
estimated values according to the curve
}
\description{
svi original function
}
