% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_funs.R
\name{u_bs}
\alias{u_bs}
\title{BS total IV approximation (upper bound)}
\usage{
u_bs(c, k, cp = 1)
}
\arguments{
\item{c}{normalized option price C/D/F (call), or P/D/F (put). D = e^(-rt) discount, F forward price}

\item{k}{log(K/F)}

\item{cp}{call/put indicator, call = 1, put = -1}
}
\description{
The upper bound of total IV, given by Gatheral et al., 2017
}
