% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_funs.R
\name{UR.interp}
\alias{UR.interp}
\title{interpolate U and R}
\usage{
UR.interp(param, tau_interp, price)
}
\arguments{
\item{param}{SVI parameter output table, including U, R etc.}

\item{tau_interp}{tau of interpolation}

\item{price}{Spot price}
}
\value{
data.frame of interpolated U, R, mean Forward Price mF, tau, price
}
\description{
interpolate U and R
}
