% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_funs.R
\name{vs.deltaStrikes}
\alias{vs.deltaStrikes}
\title{Delta Strikes}
\usage{
vs.deltaStrikes(deltas, surf, tau, callput = "p", minGap = 5,
  strikeDelta = FALSE)
}
\arguments{
\item{deltas}{input deltas (either BS delta or true delta)}

\item{surf}{surf parameters}

\item{tau}{maturity}

\item{callput}{call/put, default 'p'}

\item{minGap}{minimum interval in the option chain, by default is 5}

\item{strikeDelta}{logical option for use strikeDelta}
}
\description{
From Deltas to calculate the strikes from a surface and a given tau.
Note, by default the input Delta is a normal BS delta. It is also possible to
calculate strike-delta (delta - vega * skew slope) by set strikeDelta = TRUE.
by default, input delta are positive (even for Put side)
}
