% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_calibration.R
\name{vs.fitSurface}
\alias{vs.fitSurface}
\title{Fit Vol Surface with the transform method}
\usage{
vs.fitSurface(ochain, parallel = TRUE)
}
\arguments{
\item{ochain}{input chain}
}
\value{
estimated parameters
}
\description{
Estimate surface by transforming and translating the surface to a uniform surface
Using this to find rho and H. The whole surface uses the same rho and H.
Here we assume the curvation of the iv against k/sqrt(tau) is fixed (H), and
the smile rotation is fixed as well
}
