% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calibration.R
\name{vs.greeks}
\alias{vs.greeks}
\title{Obtain greeks from vertical skew}
\usage{
vs.greeks(surf, tau, callput = "p", k = seq(-0.5, 0.3, length.out = 1200))
}
\description{
Obtain greeks from a surface with a given tau
Rework on vs.options with array input
}
