% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_calibration.R
\name{vs.interParameter}
\alias{vs.interParameter}
\title{interpolation of parameters}
\usage{
vs.interParameter(surf, tau_interp)
}
\arguments{
\item{surf}{surface parameters}

\item{tau_interp}{to-be-interpolated tau}
}
\value{
interpolated parameters
Note: Interpolation works only within the maturity range.
To interpolate the parameters, one has to convert raw parameters to
jump wing parameters, given that jump wing parameters have clear meanings.
That is, iv0, skew, left/right slopes, minimum. Those parameters can be
easily interpolated.
}
\description{
interpolation of parameters
}
