% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_calibration.R
\name{vs.ivs}
\alias{vs.ivs}
\title{Estimate iv from trans.method parameters}
\usage{
vs.ivs(k, para)
}
\arguments{
\item{k}{log moneyness}

\item{para}{parameters include maturity, iv0, a,b,m, rho, sigma H, Ht, Sx,Sy}
}
\description{
Estimate iv from trans.method parameters
}
