% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calibration.R
\name{vs.options}
\alias{vs.options}
\title{estimate price, skew and nrd at a given DTE}
\usage{
vs.options(surface, dte, spot = NULL, type = "put", x = seq(-0.2, 0.1,
  length.out = 600))
}
\arguments{
\item{surface}{a surface param, estimated from SVI}

\item{dte}{Days to expiration}

\item{x}{a sequence of logmoneyness or strike. If max(x)<1, it will be regarded as log moneyness.}
}
\description{
This function will estimate options prices etc at a given iv
curve with a specified dte
}
