% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_calibration.R
\name{vs.plotSurface}
\alias{vs.plotSurface}
\title{plot iv surface with fitted model}
\usage{
vs.plotSurface(ivchain, para, range = c(10, 90), type = "iv",
  x = "strike")
}
\arguments{
\item{ivchain}{iv chain}

\item{para}{svi parameters}

\item{range}{dte range (default c(7,90))}

\item{type}{default ('iv') plot normal IV figure, or 'ivt', plot iv/sqrt(t), or 'var', plot total}

\item{x}{default ('strike'), alternative 'logmoneyness'}
}
\description{
plot iv surface with fitted model
}
