% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calibration.R
\name{vs.termstructure}
\alias{vs.termstructure}
\title{Plot term structure}
\usage{
vs.termstructure(surface)
}
\arguments{
\item{surface}{surface parameters}
}
\description{
Based on modified svi surface, estimate ATM IVs across DTEs
}
