% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calibration.R
\name{vs.vskew}
\alias{vs.vskew}
\title{Esimate Vertical Skew}
\usage{
vs.vskew(x, surf, dte, spot = NULL)
}
\arguments{
\item{x}{Strikes or log(K/F)}

\item{surf}{surface parameters}

\item{dte}{Day-to-expiration}

\item{spot}{spot value for sticky-delta calculation

When the spot is not specify, vertical estimation is soly dependent
on the surf, so it is sticky-strike. When the spot is specify,
the surface will move and allocate to the spot (i.e., sticky-delta)}
}
\description{
Esimate vertical skew curve with given x, surface, and dte
}
