% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funs.R
\name{vue.iv}
\alias{vue.iv}
\title{estimate iv from optionVue chain}
\usage{
vue.iv(filename, LR)
}
\arguments{
\item{filename}{optionVue chain}

\item{LR}{libor rate}
}
\value{
refined chain with estimated IV
}
\description{
estimate iv from optionVue chain
}
